Resolved window not ready bug, this caused rbd to break.
