+++ package.json
@@ -8,7 +8,10 @@
     "graphql": "node graphql/index.js"
   },
   "dependencies": {
+    "@feizheng/next-local-storage": "^1.1.0",
     "@headlessui/react": "^1.2.0",
+    "@jswork/next": "^1.0.7",
+    "@jswork/next-local-storage": "^1.0.0",
     "@material-ui/core": "^4.11.4",
     "animate.css": "^4.1.1",
     "aos": "^2.3.4",
@@ -21,6 +24,7 @@
     "firebase": "^8.6.5",
     "graphql": "^15.5.0",
     "html-metadata-parser": "^1.0.4",
+    "local-storage": "^2.0.0",
     "lodash": "^4.17.21",
     "next": "^11.0.0",
     "next-pwa": "^5.2.21",
